ISG_agent/requirements_agent.txt
absl-py==2.1.0
accelerate==0.34.2
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
albucore==0.0.15
albumentations==1.4.15
aliyun-python-sdk-core==2.15.2
aliyun-python-sdk-kms==2.16.5
altair==5.4.1
annotated-types==0.7.0
anthropic==0.34.2
antlr4-python3-runtime==4.9.3
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrdict==2.0.1
attrs==24.2.0
autocommand==2.2.2
av==13.0.0
babel==2.16.0
backports.tarfile==1.2.0
beartype==0.18.5
beautifulsoup4==4.12.3
bidict==0.23.1
binpacking==1.5.2
black==24.2.0
bleach==6.1.0
blinker==1.8.2
boto3==1.35.17
botocore==1.35.17
braceexpand==0.1.7
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.1.0
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpickle==3.0.0
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
cpm-kernels==1.0.11
crcmod==1.7
cryptography==43.0.1
cupy-cuda12x==13.3.0
cycler==0.12.1
dacite==1.8.1
datasets==3.0.0
debugpy==1.8.5
decorator==4.4.2
decord==0.6.0
deepspeed==0.14.4
defusedxml==0.7.1
-e git+ssh://git@github.com/HaozheZhao/UltraEdit.git@1af5f0478d56d62cf6f35781a2ab63e03d656b2d#egg=diffusers&subdirectory=diffusers
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.0
einx==0.3.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
executing==2.1.0
fairscale==0.4.13
fastapi==0.114.1
fastjsonschema==2.20.0
fastrlock==0.8.2
ffmpy==0.4.0
filelock==3.16.0
fire==0.6.0
flash-attn==2.5.8
Flask==3.0.3
Flask-SocketIO==5.3.7
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.3
future==1.0.0
fvcore==0.1.5.post20221221
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.9
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.0
googleapis-common-protos==1.65.0
gradio==4.44.0
gradio_client==1.3.0
grpcio==1.66.1
grpcio-status==1.48.2
h11==0.14.0
hjson==3.1.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.24.7
humanfriendly==10.0
hydra-core==1.3.2
idna==3.8
imageio==2.34.2
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
importlib_resources==6.4.0
inflect==7.3.1
invisible-watermark==0.2.0
iopath==0.1.10
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
jaxtyping
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==0.10.0
joblib==1.4.2
json5==0.9.25
jsonfy==0.4
jsonify==0.5
jsonlines==4.0.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
kornia==0.7.3
kornia_rs==0.1.5
lazy_loader==0.4
libcst==1.4.0
lightning-utilities==0.11.7
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
modelscope==1.18.0
more-itertools==10.3.0
moreorless==0.4.0
moviepy==1.0.3
mpi4py_mpich==3.1.5
mpmath==1.3.0
ms-swift==2.4.1
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.7.0
natsort==8.4.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnxruntime==1.19.2
open-clip-torch==2.22.0
openai==1.44.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
orjson==3.10.7
oss2==2.19.0
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
peft==0.12.0
pexpect==4.9.0
Pillow==10.0.1
platformdirs==4.3.2
pooch==1.8.2
portalocker==2.10.1
proglog==0.1.10
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pudb==2024.1.2
pure_eval==0.2.3
py==1.11.0
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.1
pydantic_core==2.23.3
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
PyMatting==1.1.12
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-json-logger==2.0.7
python-multipart==0.0.9
python-socketio==5.11.4
pytorch-lightning==2.4.0
pytz==2024.2
PyWavelets==1.7.0
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.9.11
rembg==2.0.59
replicate==0.33.0
requests==2.32.3
retry==0.9.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.1
rotary-embedding-torch==0.8.3
rouge==1.0.1
rpds-py==0.20.0
rsa==4.9
ruff==0.6.4
s3transfer==0.10.2
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==2.14.0
setproctitle==1.3.3
share==1.0.4
shellingham==1.5.4
shtab==1.7.1
simple-websocket==1.0.0
simplejson==3.19.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spaces==0.30.2
stack-data==0.6.3
starlette==0.38.5
stdlibs==2024.5.15
streamlit==1.38.0
streamlit-keyup==0.2.0
submitit==1.5.1
supervision==0.23.0
SwissArmyTransformer==0.4.12
sympy==1.13.2
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorboardX==2.6
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.8.30
timm==1.0.9
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
torch==2.3.1
torchaudio==2.3.1
torchdata==0.8.0
torchdiffeq==0.2.4
torchmetrics==1.4.1
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.5
trailrunner==1.4.0
traitlets==5.14.3
transformers==4.44.2
transformers-stream-generator==0.0.5
transfusion-pytorch==0.0.42
triton==2.3.1
trl==0.10.1
typeguard==4.3.0
typer==0.12.5
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
tyro==0.8.10
tzdata==2024.1
ufmt==2.0.0b2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.0.0
urwid==2.6.15
urwid_readline==0.14
usort==1.0.2
uvicorn==0.30.6
wandb==0.18.0
watchdog==4.0.2
wcwidth==0.2.13
webcolors==24.8.0
webdataset==0.2.100
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.4
widgetsnbextension==4.0.13
wsproto==1.2.0
xformers==0.0.27
xxhash==3.5.0
yacs==0.1.8
yarl==1.11.1
zipp==3.20.1
